+++ src/cpuid/src/template/mod.rs
@@ -3,13 +3,15 @@
 
 // Contains Intel specific templates.
 pub mod intel;
+// Contains AMD specific templates.
+pub mod amd;
 
 use std::collections::HashSet;
 
 use arch_gen::x86::msr_index::*;
 use kvm_bindings::CpuId;
 
-use crate::common::{get_vendor_id_from_cpuid, VENDOR_ID_INTEL};
+use crate::common::{get_vendor_id_from_cpuid, VENDOR_ID_AMD, VENDOR_ID_INTEL};
 use crate::cpuid_is_feature_set;
 use crate::transformer::Error;
 
@@ -77,7 +79,7 @@ fn cpuid_to_msrs_to_save(cpuid: &CpuId) -> HashSet<u32> {
 pub fn msrs_to_save_by_cpuid(cpuid: &CpuId) -> Result<HashSet<u32>, Error> {
     let vendor_id = get_vendor_id_from_cpuid(cpuid).map_err(|_| Error::InvalidVendor)?;
     match &vendor_id {
-        VENDOR_ID_INTEL => Ok(cpuid_to_msrs_to_save(cpuid)),
+        VENDOR_ID_INTEL | VENDOR_ID_AMD => Ok(cpuid_to_msrs_to_save(cpuid)),
         _ => {
             // We don't have MSR-CPUID dependencies set for other vendors yet.
             Ok(HashSet::new())