+++ tests/integration_tests/functional/test_feat_parity.py
@@ -10,10 +10,11 @@ from framework.artifacts import DiskArtifact
 from framework.builder import MicrovmBuilder
 import framework.utils_cpuid as cpuid_utils
 from framework.utils_cpu_templates import SUPPORTED_CPU_TEMPLATES
+import host_tools.network as net_tools
 
 
 # CPU templates designed to provide instruction set feature parity
-INST_SET_TEMPLATES = []
+INST_SET_TEMPLATES = ["T2A", "T2CL"]
 
 
 @pytest.fixture(name="vm_builder", scope="session")
@@ -44,6 +45,15 @@ def disk_fxt(request):
     disk = request.param
     disk.download()
     return disk
+
+
+@pytest.fixture(
+    name="inst_set_cpu_template",
+    params=set(SUPPORTED_CPU_TEMPLATES).intersection(INST_SET_TEMPLATES),
+)
+def inst_set_cpu_template_fxt(request):
+    """CPU template fixture for instruction set feature parity templates"""
+    return request.param
 
 
 @pytest.fixture(
@@ -209,3 +219,126 @@ def test_feat_parity_cpuid_inst_set(
         must_be_set,
         must_be_unset,
     )
+
+
+def test_feat_parity_cpuid_sec(
+    vm_builder, inst_set_cpu_template, microvm, guest_kernel, disk
+):
+    """
+    Verify that security-related CPUID feature flags are properly set
+    for T2CL and T2A CPU templates.
+
+    @type: functional
+    """
+
+    # fmt: off
+    must_be_set_common = [
+        (0x7, 0x0, "edx",
+            (1 << 26) | # IBRS/IBPB
+            (1 << 27) | # STIBP
+            (1 << 31) # SSBD
+        )
+        # Security feature bits in 0x80000008 EBX are set differently by
+        # 4.14 and 5.10 KVMs.
+        # 4.14 populates them from host's AMD flags (0x80000008 EBX), while
+        # 5.10 takes them from host's common flags (0x7 EDX).
+        # There is no great value in checking that this actually happens, as
+        # we cannot really control it.
+        # When we drop 4.14 support, we may consider enabling this check.
+        # (0x80000008, 0x0, "ebx",
+        #     (1 << 12) | # IBPB
+        #     (1 << 14) | # IBRS
+        #     (1 << 15) | # STIBP
+        #     (1 << 24) # SSBD
+        # )
+    ]
+
+    must_be_set_intel_only = [
+        (0x7, 0x0, "edx",
+            (1 << 10) | # MD_CLEAR
+            (1 << 29) # IA32_ARCH_CAPABILITIES
+        )
+    ]
+
+    must_be_set_amd_only = [
+        (0x80000008, 0x0, "ebx",
+            (1 << 18) | # IbrsPreferred
+            (1 << 19) # IbrsProvidesSameModeProtection
+        )
+    ]
+
+    must_be_unset_common = [
+        (0x7, 0x0, "edx",
+            (1 << 28) # L1D_FLUSH
+        )
+    ]
+
+    must_be_unset_intel_only = [
+        (0x80000008, 0x0, "ebx",
+            (1 << 18) | # IbrsPreferred
+            (1 << 19) # IbrsProvidesSameModeProtection
+        )
+    ]
+
+    must_be_unset_amd_only = [
+        (0x7, 0x0, "edx",
+            (1 << 10) | # MD_CLEAR
+            (1 << 29) # IA32_ARCH_CAPABILITIES
+        )
+    ]
+    # fmt: on
+
+    vendor = cpuid_utils.get_cpu_vendor()
+    if vendor == cpuid_utils.CpuVendor.INTEL:
+        must_be_set = must_be_set_common + must_be_set_intel_only
+        must_be_unset = must_be_unset_common + must_be_unset_intel_only
+    elif vendor == cpuid_utils.CpuVendor.AMD:
+        must_be_set = must_be_set_common + must_be_set_amd_only
+        must_be_unset = must_be_unset_common + must_be_unset_amd_only
+    else:
+        raise Exception("Unsupported CPU vendor.")
+
+    check_cpuid_feat_flags(
+        vm_builder,
+        inst_set_cpu_template,
+        microvm,
+        guest_kernel,
+        disk,
+        must_be_set,
+        must_be_unset,
+    )
+
+
+def test_feat_parity_msr_arch_cap(
+    vm_builder, inst_set_cpu_template, microvm, guest_kernel, disk
+):
+    """
+    Verify availability and value of the IA32_ARCH_CAPABILITIES MSR for T2CL and T2A CPU templates.
+
+    @type: functional
+    """
+    vm = create_vm(vm_builder, inst_set_cpu_template, microvm, guest_kernel, disk)
+    vm.start()
+
+    ssh_conn = net_tools.SSHConnection(vm.ssh_config)
+    arch_capabilities_addr = "0x10a"
+    rdmsr_cmd = f"rdmsr {arch_capabilities_addr}"
+    _, stdout, stderr = ssh_conn.execute_command(rdmsr_cmd)
+
+    if inst_set_cpu_template == "T2CL":
+        assert stderr.read() == ""
+        actual = int(stdout.read().strip(), 16)
+        # fmt: off
+        expected = (
+            (1 << 0) | # RDCL_NO
+            (1 << 1) | # IBRS_ALL
+            (1 << 3) | # SKIP_L1DFL_VMENTRY
+            (1 << 5) | # MDS_NO
+            (1 << 6) | # IF_PSCHANGE_MC_NO
+            (1 << 7) # TSX_CTRL
+        )
+        # fmt: on
+        assert actual == expected, f"{actual=:#x} != {expected=:#x}"
+    elif inst_set_cpu_template == "T2A":
+        # IA32_ARCH_CAPABILITIES shall not be available
+        assert stderr.read() != ""